Rename sub directory to AlgorithmRunners
