+++ qt/scientific_interfaces/Indirect/IndirectDataReduction.h
@@ -12,7 +12,7 @@
 
 #include "MantidAPI/MatrixWorkspace.h"
 #include "MantidGeometry/IComponent.h"
-#include "MantidQtWidgets/Common/JobRunners/AlgorithmRunner.h"
+#include "MantidQtWidgets/Common/AlgorithmRunners/AlgorithmRunner.h"
 
 #include <QRegExp>
 #include <QScrollArea>
