+++ qt/scientific_interfaces/Indirect/IndirectSimulationTab.h
@@ -11,7 +11,7 @@
 
 #include "MantidAPI/AlgorithmManager.h"
 #include "MantidKernel/System.h"
-#include "MantidQtWidgets/Common/JobRunners/AlgorithmRunner.h"
+#include "MantidQtWidgets/Common/AlgorithmRunners/AlgorithmRunner.h"
 
 #include <QSettings>
 #include <QWidget>
