+++ qt/widgets/common/inc/MantidQtWidgets/Common/AlgorithmRunners/BatchAlgorithmRunner.h
