+++ qt/widgets/common/inc/MantidQtWidgets/Common/AlgorithmRunners/IJobRunner.h
