+++ qt/widgets/common/inc/MantidQtWidgets/Common/AlgorithmRunners/MockJobRunner.h
@@ -7,7 +7,7 @@
 #pragma once
 
 #include "MantidKernel/WarningSuppressions.h"
-#include "MantidQtWidgets/Common/JobRunners/IJobRunner.h"
+#include "MantidQtWidgets/Common/AlgorithmRunners/IJobRunner.h"
 
 #include <gmock/gmock.h>
 #include <memory>
